Move uisettings-related API tests
